rename: 데이터 파일, dom.js 파일 경로 변경
